coded greet.js and test.js ... tests failing ... debug in progress
